+++ tests/it/main.rs
@@ -1,3 +1,4 @@
 mod alloc;
 mod bitmap;
 mod buffer;
+mod ffi;
